refactor out new branching code and re-use it for calculation termination conditions
