+++ src/Griddly/Core/GDY/Objects/Object.cpp
@@ -137,27 +137,49 @@ BehaviourCondition Object::resolveConditionArguments(const std::function<bool(in
   };
 }
 
-BehaviourCondition Object::instantiateCondition(std::string &commandName, YAML::Node &conditionNode) const {
-  if (commandName == "eq") {
-    return resolveConditionArguments([](int32_t a, int32_t b) { return a == b; }, conditionNode);
-  } else if (commandName == "gt") {
-    return resolveConditionArguments([](int32_t a, int32_t b) { return a > b; }, conditionNode);
-  } else if (commandName == "gte") {
-    return resolveConditionArguments([](int32_t a, int32_t b) { return a >= b; }, conditionNode);
-  } else if (commandName == "lt") {
-    return resolveConditionArguments([](int32_t a, int32_t b) { return a < b; }, conditionNode);
-  } else if (commandName == "lte") {
-    return resolveConditionArguments([](int32_t a, int32_t b) { return a <= b; }, conditionNode);
-  } else if (commandName == "neq") {
-    return resolveConditionArguments([](int32_t a, int32_t b) { return a != b; }, conditionNode);
-  } else if (commandName == "and") {
-    return processConditions(conditionNode, false, LogicOp::AND);
-  } else if (commandName == "or") {
-    return processConditions(conditionNode, false, LogicOp::OR);
-  } else {
-    throw std::invalid_argument(fmt::format("Unknown or badly defined condition command {0}.", commandName));
-  }
+BehaviourCondition Object::resolveAND(const std::vector<BehaviourCondition>& conditionList) const {
+  return [conditionList](const std::shared_ptr<Action> &action) -> bool {
+    for (const auto &condition : conditionList) {
+      if (!condition(action)) {
+        return false;
+      }
+    }
+    return true;
+  };
 }
+
+BehaviourCondition Object::resolveOR(const std::vector<BehaviourCondition>& conditionList) const {
+  return [conditionList](const std::shared_ptr<Action> &action) -> bool {
+    for (const auto &condition : conditionList) {
+      if (condition(action)) {
+        return true;
+      }
+    }
+    return false;
+  };
+}
+
+// BehaviourCondition Object::instantiateCondition(std::string &commandName, YAML::Node &conditionNode) const {
+//   if (commandName == "eq") {
+//     return resolveConditionArguments([](int32_t a, int32_t b) { return a == b; }, conditionNode);
+//   } else if (commandName == "gt") {
+//     return resolveConditionArguments([](int32_t a, int32_t b) { return a > b; }, conditionNode);
+//   } else if (commandName == "gte") {
+//     return resolveConditionArguments([](int32_t a, int32_t b) { return a >= b; }, conditionNode);
+//   } else if (commandName == "lt") {
+//     return resolveConditionArguments([](int32_t a, int32_t b) { return a < b; }, conditionNode);
+//   } else if (commandName == "lte") {
+//     return resolveConditionArguments([](int32_t a, int32_t b) { return a <= b; }, conditionNode);
+//   } else if (commandName == "neq") {
+//     return resolveConditionArguments([](int32_t a, int32_t b) { return a != b; }, conditionNode);
+//   } else if (commandName == "and") {
+//     return processConditions(conditionNode, false, LogicOp::AND);
+//   } else if (commandName == "or") {
+//     return processConditions(conditionNode, false, LogicOp::OR);
+//   } else {
+//     throw std::invalid_argument(fmt::format("Unknown or badly defined condition command {0}.", commandName));
+//   }
+// }
 
 BehaviourFunction Object::instantiateConditionalBehaviour(std::string &commandName, BehaviourCommandArguments &commandArguments, CommandList &subCommands) {
   if (subCommands.size() == 0) {
@@ -204,76 +226,64 @@ BehaviourFunction Object::instantiateConditionalBehaviour(std::string &commandNa
   };
 }
 
-/**
- * @brief Deliciously recursive function for processing trees of conditions
- * 
- * @param conditionNodeList 
- * @param isTopLevel 
- * @return BehaviourCondition 
- */
-BehaviourCondition Object::processConditions(YAML::Node &conditionNodeList, bool isTopLevel, LogicOp op) const {
-  // We should have a single item and not a list
-  if (!conditionNodeList.IsDefined()) {
-    auto line = conditionNodeList.Mark().line;
-    auto errorString = fmt::format("Parse error line {0}. If statement is missing Conditions", line);
-    spdlog::error(errorString);
-    throw std::invalid_argument(errorString);
-  }
+// BehaviourCondition Object::processConditions(YAML::Node &conditionNodeList, bool isTopLevel, LogicOp op) const {
+//   // We should have a single item and not a list
+//   if (!conditionNodeList.IsDefined()) {
+//     auto line = conditionNodeList.Mark().line;
+//     auto errorString = fmt::format("Parse error line {0}. If statement is missing Conditions", line);
+//     spdlog::error(errorString);
+//     throw std::invalid_argument(errorString);
+//   }
 
-  if (conditionNodeList.IsMap()) {
-    if (conditionNodeList.size() != 1) {
-      auto line = conditionNodeList.Mark().line;
-      auto errorString = fmt::format("Parse error line {0}. Conditions must contain a single top-level condition", line);
-      spdlog::error(errorString);
-      throw std::invalid_argument(errorString);
-    }
-    auto conditionNode = conditionNodeList.begin();
-    auto commandName = conditionNode->first.as<std::string>();
-    return instantiateCondition(commandName, conditionNode->second);
+//   if (conditionNodeList.IsMap()) {
+//     if (conditionNodeList.size() != 1) {
+//       auto line = conditionNodeList.Mark().line;
+//       auto errorString = fmt::format("Parse error line {0}. Conditions must contain a single top-level condition", line);
+//       spdlog::error(errorString);
+//       throw std::invalid_argument(errorString);
+//     }
+//     auto conditionNode = conditionNodeList.begin();
+//     auto commandName = conditionNode->first.as<std::string>();
+//     return instantiateCondition(commandName, conditionNode->second);
 
-  } else if (conditionNodeList.IsSequence()) {
-    std::vector<BehaviourCondition> conditionList;
-    for (auto &&subConditionNode : conditionNodeList) {
-      auto validatedNode = validateCommandPairNode(subConditionNode);
-      auto commandName = validatedNode->first.as<std::string>();
-      conditionList.push_back(instantiateCondition(commandName, validatedNode->second));
-    }
-    switch (op) {
-      case LogicOp::AND: {
-        return [conditionList](const std::shared_ptr<Action> &action) -> bool {
-          for (const auto &condition : conditionList) {
-            if (!condition(action)) {
-              return false;
-            }
-          }
-          return true;
-        };
-      } break;
-      case LogicOp::OR: {
-        return [conditionList](const std::shared_ptr<Action> &action) -> bool {
-          for (const auto &condition : conditionList) {
-            if (condition(action)) {
-              return true;
-            }
-          }
-          return false;
-        };
-      }
-      default: {
-        auto line = conditionNodeList.Mark().line;
-        auto errorString = fmt::format("Parse error line {0}. A sequence of conditions must be within an AND or an OR operator.", line);
-        spdlog::error(errorString);
-        throw std::invalid_argument(errorString);
-      }
-    }
-
-  } else {
-    auto line = conditionNodeList.Mark().line;
-    auto errorString = fmt::format("Conditions must be a map or a list", line);
-    spdlog::error(errorString);
-    throw std::invalid_argument(errorString);
-  }
-}
+//   } else if (conditionNodeList.IsSequence()) {
+//     std::vector<BehaviourCondition> conditionList;
+//     for (auto &&subConditionNode : conditionNodeList) {
+//       auto validatedNode = validateCommandPairNode(subConditionNode);
+//       auto commandName = validatedNode->first.as<std::string>();
+//       conditionList.push_back(instantiateCondition(commandName, validatedNode->second));
+//     }
+//     switch (op) {
+//       case LogicOp::AND: {
+//         return resolveAnd()
+//       };
+//     }
+//     break;
+//     case LogicOp::OR: {
+//       return [conditionList](const std::shared_ptr<Action> &action) -> bool {
+//         for (const auto &condition : conditionList) {
+//           if (condition(action)) {
+//             return true;
+//           }
+//         }
+//         return false;
+//       };
+//     }
+//     default: {
+//       auto line = conditionNodeList.Mark().line;
+//       auto errorString = fmt::format("Parse error line {0}. A sequence of conditions must be within an AND or an OR operator.", line);
+//       spdlog::error(errorString);
+//       throw std::invalid_argument(errorString);
+//     }
+//   }
+// }
+// else {
+//   auto line = conditionNodeList.Mark().line;
+//   auto errorString = fmt::format("Conditions must be a map or a list", line);
+//   spdlog::error(errorString);
+//   throw std::invalid_argument(errorString);
+// }
+// }  // namespace griddly
 
 /**
  * @brief executes a list of behaviour functions and accumulates the rewards
